updated launch_ahsm.sh

+++ checkstyle-tester/launch_ahsm.sh
@@ -6,8 +6,8 @@ set -e
 # Note: Use full paths
 # ============================================================
 
-MINIMIZE=false
-CONTACTSERVER=false
+MINIMIZE=true
+CONTACTSERVER=true
 
 CHECKSTYLE_DIR=~/opensource/checkstyle
 TESTER_DIR=~/opensource/contribution/checkstyle-tester
@@ -19,9 +19,9 @@ AHSM_JAR=~/Downloads/ahsm.jar
 
 PULL_REMOTE=pull
 
-SITE_SOURCES_DIR_NAME=src/main/java
-SITE_SAVE_MASTER_DIR_NAME=savemaster
-SITE_SAVE_PULL_DIR_NAME=savepull
+SITE_SOURCES_DIR=src/main/java
+SITE_SAVE_MASTER_DIR=savemaster
+SITE_SAVE_PULL_DIR=savepull
 
 # ============================================================
 # ============================================================
@@ -29,6 +29,50 @@ SITE_SAVE_PULL_DIR_NAME=savepull
 
 #declare -a EXTPROJECTS
 EXTPROJECTS=()
+INSTALL_MASTER=true
+
+if [ -z "$1" ]; then
+	echo "No parameter supplied!"
+	echo "      Either supply the PR branch to work with or 'clean'."
+	echo ""
+	exit 1
+fi
+
+if [ "$1" == clean ]; then
+	echo "Cleaning..."
+
+	cd $CHECKSTYLE_DIR
+	mvn --batch-mode clean
+	cd $TESTER_DIR
+	rm -rf $SITE_SOURCES_DIR/*
+	rm -rf $SITE_SAVE_MASTER_DIR
+	rm -rf $SITE_SAVE_PULL_DIR
+	mvn --batch-mode clean
+	rm -rf $FINAL_RESULTS_DIR/*
+	exit 0
+fi
+
+function parse_arguments {
+	SKIP=true
+
+	while [[ $@ > 0 ]] ; do
+		if $SKIP ; then
+			case "$1" in
+			-skip)
+				case "$2" in
+				install_master)
+					INSTALL_MASTER=false
+					;;
+				esac
+				shift
+				;;
+			esac
+		else
+			SKIP=false
+		fi
+		shift
+	done
+}
 
 function mvn_install {
 	mvn --batch-mode clean install -Dmaven.test.skip=true -Dcheckstyle.ant.skip=true
@@ -43,7 +87,7 @@ function launch {
 		cd $TESTER_DIR
 
 		while read line ; do
-			rm -rf $SITE_SOURCES_DIR_NAME/*
+			rm -rf $SITE_SOURCES_DIR/*
 			
 			[[ "$line" == \#* ]] && continue # Skip lines with comments
 			[[ -z "$line" ]] && continue     # Skip empty lines
@@ -56,7 +100,7 @@ function launch {
 			
 			echo "Running Launch on $REPO_NAME ..."
 			
-			REPO_SOURCES_DIR=$SITE_SOURCES_DIR_NAME/$REPO_NAME
+			REPO_SOURCES_DIR=$SITE_SOURCES_DIR/$REPO_NAME
 			
 			if [ ! -d "$EXTRA_DIR" ]; then
 				mkdir $EXTRA_DIR
@@ -72,7 +116,7 @@ function launch {
 				fi
 				if [ ! -d "$REPO_SOURCES_DIR" ]; then
 					echo -e "untar $TARNAME file to $REPO_SOURCES_DIR ..."
-					tar -xf $TARPATH -C $SITE_SOURCES_DIR_NAME
+					tar -xf $TARPATH -C $SITE_SOURCES_DIR
 					echo -e "untar $TARNAME file to $REPO_SOURCES_DIR - completed"
 				fi
 			elif [ "$REPO_TYPE" == "git" ]; then
@@ -99,14 +143,14 @@ function launch {
 				exit 1
 			fi
 
-			echo "Running Checkstyle on $SITE_SOURCES_DIR_NAME ... with excludes $EXCLUDES_ACCUM"
+			echo "Running Checkstyle on $SITE_SOURCES_DIR ... with excludes $EXCLUDES_ACCUM"
 			mvn -e --batch-mode clean site -Dcheckstyle.excludes=$EXCLUDES -DMAVEN_OPTS=-Xmx3024m
 			if [ "$?" != "0" ]
 			then
-				echo "Checkstyle failed on $SITE_SOURCES_DIR_NAME"
+				echo "Checkstyle failed on $SITE_SOURCES_DIR"
 				exit 1
 			else
-				echo "Running Checkstyle on $SITE_SOURCES_DIR_NAME - finished"
+				echo "Running Checkstyle on $SITE_SOURCES_DIR - finished"
 			fi
 
 			echo "linking report to index.html"
@@ -128,7 +172,12 @@ function launch {
 
 			echo "Backuping files that are refenced in report ..."
 			for f in $(cat file.txt) ; do
-			  mv "$f" "$f.save"
+				if [ -f "$f" ]
+				then
+					mv "$f" "$f.save"
+				else
+					echo "warning: $f not found."
+				fi
 			done
 
 			echo "Removing all non used html files"
@@ -136,7 +185,12 @@ function launch {
 
 			echo "Restoring from backup.."
 			for f in $(cat file.txt) ; do
-			  mv "$f.save" "$f"
+				if [ -f "$f.save" ]
+				then
+					mv "$f.save" "$f"
+				else
+					echo "warning: $f.save not found."
+				fi
 			done
 
 			VIOLATIONS=$(grep '<th>Line</th>' ../index.html | wc -l)
@@ -168,41 +222,49 @@ function launch {
 # ============================================================
 # ============================================================
 
+parse_arguments "$@"
+
 echo "Testing Checkstyle Starting"
 
-cd $CHECKSTYLE_DIR
+if $INSTALL_MASTER ; then
+	cd $CHECKSTYLE_DIR
 
-if $CONTACTSERVER ; then
-	git fetch origin
-	git fetch $PULL_REMOTE
+	if $CONTACTSERVER ; then
+		git fetch origin
+		git fetch $PULL_REMOTE
+	fi
+	git reset --hard HEAD
+	git checkout master
+	git pull
+
+	echo "Installing Master"
+
+	mvn_install
+else
+	echo "Skipping Install Master"
 fi
-git reset --hard HEAD
-git checkout master
-
-echo "Installing Master"
-
-mvn_install
 
 echo "Starting Master Launcher"
 
 cd $TESTER_DIR
 
-rm -rf $SITE_SAVE_MASTER_DIR_NAME
-rm -rf $SITE_SAVE_PULL_DIR_NAME
+rm -rf $SITE_SAVE_MASTER_DIR
+rm -rf $SITE_SAVE_PULL_DIR
 
-launch $SITE_SAVE_MASTER_DIR_NAME
+launch $SITE_SAVE_MASTER_DIR
 
 cd $CHECKSTYLE_DIR
 
 echo "Checking out and Installing PR $1"
 
 git checkout $1
+git pull
 
 mvn_install
 
 echo "Starting PR $1 Launcher"
 
-launch $SITE_SAVE_PULL_DIR_NAME
+launch $SITE_SAVE_PULL_DIR
 
 echo "Starting all AHSMs"
 
@@ -213,9 +275,9 @@ fi
 for extp in "${EXTPROJECTS[@]}"
 do
 	if [ ! -d "$FINAL_RESULTS_DIR/$extp" ]; then
-		echo "java -jar $AHSM_JAR checkstyle $TESTER_DIR/$SITE_SAVE_MASTER_DIR_NAME/$extp $TESTER_DIR/$SITE_SAVE_PULL_DIR_NAME/$extp $FINAL_RESULTS_DIR/$extp"
+		echo "java -jar $AHSM_JAR checkstyle $TESTER_DIR/$SITE_SAVE_MASTER_DIR/$extp $TESTER_DIR/$SITE_SAVE_PULL_DIR/$extp $FINAL_RESULTS_DIR/$extp"
 	
-		java -jar $AHSM_JAR checkstyle $TESTER_DIR/$SITE_SAVE_MASTER_DIR_NAME/$extp $TESTER_DIR/$SITE_SAVE_PULL_DIR_NAME/$extp $FINAL_RESULTS_DIR/$extp
+		java -jar $AHSM_JAR checkstyle $TESTER_DIR/$SITE_SAVE_MASTER_DIR/$extp $TESTER_DIR/$SITE_SAVE_PULL_DIR/$extp $FINAL_RESULTS_DIR/$extp
 		
 		if [ "$?" != "0" ]
 		then
